requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
duckdb
pymysql
polars
apache-airflow-providers-mysql
pyarrow
duckdb
dbt-duckdb
airflow-provider-duckdb==0.2.0